Bola's Initial Fix

+++ src/UI/Gforce.java
@@ -2,9 +2,9 @@ package UI;
 
 public class Gforce {
 
-        public static double calculateForce(double m, float v){
+        public static double calculateForce(double m, double v){
             double g= 9.81;
-            return (float) (m *g);
+            return (float) (m * g);
         }
 
     // Code to cALc Potential Energy by bola
